+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.6.0] - 2025-05-16
+- Modified the LICENSE file to include my internet nickname
+- Added automatic LSP server installation. The new LSP will ship with each new extension version
+- Added command `Textwire: Update LSP to the latest version` which you can run to update LSP binary to the latest version
+
 ## [1.5.4] - 2025-02-02
 - Changed the activation events for the extension so that it only activates when you open a `.tw` file. This should improve the performance of the extension